# docs/returns-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Returns Resolution Agent",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/returns/resolve": {
      "post": {
        "summary": "Start or continue a returns conversation",
        "operationId": "resolveReturn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "entra": [
              "api://<api-client-id>/Returns.Resolve"
            ]
          }
        ]
      }
    },
    "/v1/returns/{thread_id}/decision": {
      "post": {
        "summary": "Approve or reject a pending refund/repair",
        "operationId": "decideReturn",
        "security": [
          {
            "entra": [
              "api://<api-client-id>/Returns.Resolve"
            ]
          }
        ],
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentResponse": {
        "properties": {
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "'completed' or 'awaiting_approval'"
          },
          "reply": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reply"
          },
          "pending_action": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Action"
          }
        },
        "type": "object",
        "required": [
          "thread_id",
          "status"
        ],
        "title": "AgentResponse"
      },
      "DecisionRequest": {
        "properties": {
          "approved": {
            "type": "boolean",
            "title": "Approved"
          },
          "reviewer": {
            "type": "string",
            "title": "Reviewer",
            "description": "Name or id of the human reviewer"
          }
        },
        "type": "object",
        "required": [
          "approved",
          "reviewer"
        ],
        "title": "DecisionRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ResolveRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Customer message, e.g. 'I want to return order ORD-1001'"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id",
            "description": "Pass back to continue a conversation"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ResolveRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "entra": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://login.microsoftonline.com/<tenant-id>/oauth2/v2.0/authorize",
            "tokenUrl": "https://login.microsoftonline.com/<tenant-id>/oauth2/v2.0/token",
            "scopes": {
              "api://<api-client-id>/Returns.Resolve": "Resolve returns"
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "https://<your-app>.azurecontainerapps.io"
    }
  ]
}